+++ common/async/abort-signals.ts
@@ -74,34 +74,3 @@ export class BatchedAbortSignals {
     }
   }
 }
-
-/**
- * A helper class which allows you to use `BatchedAbortSignals` more easily. See the description
- * of that class for more info.
- */
-export class RequestCoalescer {
-  private requestsInProgress = new Map<unknown, BatchedAbortSignals>()
-
-  async makeRequest(
-    id: unknown,
-    signal: AbortSignal | undefined,
-    requestFn: (batchedSignal: AbortSignal) => Promise<void>,
-  ): Promise<void> {
-    let batchedSignals = this.requestsInProgress.get(id)
-    if (batchedSignals && !batchedSignals.aborted) {
-      batchedSignals.add(signal)
-      return
-    }
-
-    batchedSignals = new BatchedAbortSignals(signal)
-    this.requestsInProgress.set(id, batchedSignals)
-
-    try {
-      await requestFn(batchedSignals.signal)
-    } finally {
-      if (this.requestsInProgress.get(id) === batchedSignals) {
-        this.requestsInProgress.delete(id)
-      }
-    }
-  }
-}
